Remove height from grid
Storing width and height as well as the list of cells makes it possible
to represent inconsistent grids, e.g.:
    > Grid 1 1 [False, False, False, False]

With this change the semantics of a `Grid width cells` are
defined as follows:

The width parameter determines the width of the grid. The last row can
be shorter. If there is only one row, the grid can be narrower than
`width`.

The height of the grid is determined by how many times the number of
cells can be divided into chunks of length `width`.